[Refactoring] Move files
