+++ Naenio/Utils/Extension/Navigation + Extension.swift
@@ -1,0 +1,19 @@
+//
+//  Navigation + Extension.swift
+//  Naenio
+//
+//  Created by 이영빈 on 2022/08/10.
+//
+
+import SwiftUI
+
+extension UINavigationController: UIGestureRecognizerDelegate {
+    override open func viewDidLoad() {
+        super.viewDidLoad()
+        interactivePopGestureRecognizer?.delegate = self
+    }
+
+    public func gestureRecognizerShouldBegin(_ gestureRecognizer: UIGestureRecognizer) -> Bool {
+        return viewControllers.count > 1
+    }
+}
